+++ CHANGELOG.md
@@ -51,6 +51,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - (Breaking) Now setting keys were changed and use dashes, much more common for CLIs (and also makes keys consistent with what you see when running `settings`).
 - (Breaking) Now `export` expects `--jsonc` and not `--json` to be passed, matching the output filetype.
 - (Breaking) FnCPF spec slightly changed. Starting with V5 this spec will be publicly documented.
+- (Breaking) Annoyingly (but to avoid confusion), the default package manager is now `npm`.
 - Now, if installed, Bun will be the default package manager.
 - Now Git errors should be much more properly handled and reported.
 - Now the error dump file should be more readable.